+++ DummyGuessers.cpp
@@ -4,7 +4,7 @@
 std::string RandomGuesser::guess()
 {
 	std::string r = "";
-	for (size_t i = 0; i < this->length; ++i)
+	for (uint i = 0; i < this->length; ++i)
 	{
 		char c = '0' + (rand() % 10);
 		r += c;
